# backend/src/aurevia/api/v1/voice.py
"""Browser voice sessions and call status, for signed-in users."""

from __future__ import annotations

import json
import uuid
from datetime import UTC, datetime, timedelta

from fastapi import APIRouter, Query, Request, status
from sqlalchemy import case, func, select

from aurevia.errors import ServiceUnavailableError
from aurevia.identity.dependencies import (
    AgentsDep,
    CallsDep,
    PrincipalDep,
    SessionDep,
    SettingsDep,
)
from aurevia.memory.schemas import TranscriptLine, TranscriptResponse
from aurevia.memory.service import LeadService, TranscriptService
from aurevia.providers.voice_transport import VoiceTransport
from aurevia.usage.models import UsageEvent, UsageKind
from aurevia.voice.call_auth import create_call_token
from aurevia.voice.metrics import LatencySummary, summarize
from aurevia.voice.schemas import (
    AgentResponse,
    AgentUpdate,
    CallResponse,
    CallUsage,
    LatencyResponse,
    VoiceOption,
    VoiceSessionRequest,
    VoiceSessionResponse,
)
from aurevia.voice.service import AgentService, AgentSettings, CallService
from aurevia.voice.voices import APPROVED_VOICES, DEFAULT_VOICE

router = APIRouter(tags=["voice"])

# Browser participant tokens only need to outlive joining the room.
PARTICIPANT_TOKEN_TTL_SECONDS = 15 * 60


def _transport(request: Request) -> VoiceTransport:
    transport: VoiceTransport | None = request.app.state.voice_transport
    if transport is None:
        raise ServiceUnavailableError("Voice is not configured")
    return transport


def _latency(summary: LatencySummary) -> LatencyResponse:
    return LatencyResponse(
        agent_turns=summary.agent_turns,
        measured_turns=summary.measured_turns,
        interrupted_turns=summary.interrupted_turns,
        e2e_p50_ms=summary.e2e_p50_ms,
        e2e_p95_ms=summary.e2e_p95_ms,
        llm_ttft_p50_ms=summary.llm_ttft_p50_ms,
        tts_ttfb_p50_ms=summary.tts_ttfb_p50_ms,
        end_of_turn_delay_p50_ms=summary.end_of_turn_delay_p50_ms,
        target_p50_ms=summary.target_p50_ms,
        target_p95_ms=summary.target_p95_ms,
        meets_target=summary.meets_target,
    )


def _agent_response(agent: object) -> AgentResponse:
    return AgentResponse.model_validate(agent, from_attributes=True)


@router.get("/voices", summary="Approved agent voices (female and male)")
async def list_voices(principal: PrincipalDep) -> list[VoiceOption]:
    return [
        VoiceOption(id=v.id, key=v.key, name=v.name, gender=v.gender, default=v == DEFAULT_VOICE)
        for v in APPROVED_VOICES
    ]


@router.get("/agents/default", summary="The tenant's default AI agent")
async def get_default_agent(principal: PrincipalDep, session: SessionDep) -> AgentResponse:
    agent = await AgentService(session, principal.tenant_id).default_agent()
    await session.commit()  # persists the agent if this call created it
    return _agent_response(agent)


@router.put("/agents/default", summary="Configure the default AI agent (owner/admin)")
async def update_default_agent(
    body: AgentUpdate, principal: AgentsDep, session: SessionDep
) -> AgentResponse:
    agent = await AgentService(session, principal.tenant_id).update_default_agent(
        AgentSettings(**body.model_dump()), actor=principal.user_id
    )
    return _agent_response(agent)


@router.post(
    "/voice/sessions",
    status_code=status.HTTP_201_CREATED,
    summary="Start a browser voice call with the default agent",
)
async def create_voice_session(
    request: Request,
    principal: CallsDep,
    session: SessionDep,
    settings: SettingsDep,
    body: VoiceSessionRequest | None = None,
) -> VoiceSessionResponse:
    transport = _transport(request)
    assert settings.jwt_secret is not None  # noqa: S101 - authenticated requests imply it
    lead_id = body.lead_id if body else None
    if lead_id is not None:
        await LeadService(session, principal.tenant_id).get(lead_id)  # 404 if not this tenant's
    agent = await AgentService(session, principal.tenant_id).default_agent()
    call = await CallService(session, principal.tenant_id).create_browser_call(
        agent=agent, user_id=principal.user_id, lead_id=lead_id
    )
    await session.commit()

    call_token = create_call_token(
        call_id=call.id,
        tenant_id=principal.tenant_id,
        secret=settings.jwt_secret.get_secret_value(),
        ttl_seconds=settings.call_token_ttl_seconds,
    )
    await transport.open_room_with_agent(
        room=call.room,
        agent_name=settings.livekit_agent_name,
        metadata=json.dumps({"call_id": str(call.id), "call_token": call_token}),
    )
    return VoiceSessionResponse(
        call_id=call.id,
        room=call.room,
        livekit_url=transport.public_url,
        token=transport.participant_token(
            room=call.room,
            identity=f"user-{principal.user_id}",
            name="Prospect",
            ttl_seconds=PARTICIPANT_TOKEN_TTL_SECONDS,
        ),
    )


@router.get("/voice/calls/{call_id}/transcript", summary="What was said (until it expires)")
async def get_transcript(
    call_id: uuid.UUID, principal: PrincipalDep, session: SessionDep
) -> TranscriptResponse:
    await CallService(session, principal.tenant_id).get(call_id)  # 404 if not this tenant's
    lines = await TranscriptService(session, principal.tenant_id).for_call(call_id)
    return TranscriptResponse(
        call_id=call_id,
        lines=[TranscriptLine(seq=ln.seq, speaker=ln.speaker.value, text=ln.text) for ln in lines],
    )


@router.get("/voice/calls/{call_id}", summary="Call status and usage")
async def get_call(
    call_id: uuid.UUID, principal: PrincipalDep, session: SessionDep
) -> CallResponse:
    call = await CallService(session, principal.tenant_id).get(call_id)
    is_llm = UsageEvent.kind == UsageKind.LLM
    row = (
        await session.execute(
            select(
                func.coalesce(func.sum(case((is_llm, UsageEvent.input_tokens), else_=0)), 0),
                func.coalesce(func.sum(case((is_llm, UsageEvent.output_tokens), else_=0)), 0),
                func.count().filter(is_llm),
                func.count().filter(is_llm, UsageEvent.interrupted.is_(True)),
                func.coalesce(
                    func.sum(
                        case((UsageEvent.kind == UsageKind.STT, UsageEvent.audio_seconds), else_=0)
                    ),
                    0,
                ),
                func.coalesce(
                    func.sum(
                        case((UsageEvent.kind == UsageKind.TTS, UsageEvent.characters), else_=0)
                    ),
                    0,
                ),
            ).where(UsageEvent.tenant_id == principal.tenant_id, UsageEvent.call_id == call.id)
        )
    ).one()
    latency = await summarize(session, tenant_id=principal.tenant_id, call_id=call.id)
    return CallResponse(
        id=call.id,
        channel=call.channel,
        status=call.status,
        sales_state=call.sales_state,
        turn_count=call.turn_count,
        started_at=call.started_at,
        ended_at=call.ended_at,
        end_reason=call.end_reason,
        usage=CallUsage(
            llm_input_tokens=int(row[0]),
            llm_output_tokens=int(row[1]),
            llm_turns=int(row[2]),
            llm_interrupted_turns=int(row[3]),
            stt_seconds=float(row[4]),
            tts_characters=int(row[5]),
        ),
        latency=_latency(latency),
    )


@router.get("/voice/latency", summary="Turn latency across the tenant's recent calls")
async def get_latency(
    principal: PrincipalDep,
    session: SessionDep,
    days: int = Query(default=7, ge=1, le=90),
) -> LatencyResponse:
    since = datetime.now(UTC) - timedelta(days=days)
    return _latency(await summarize(session, tenant_id=principal.tenant_id, since=since))