# web/dashboard/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Aurevia API",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/v1/voices": {
   "get": {
    "tags": [
     "voice"
    ],
    "summary": "Approved agent voices (female and male)",
    "operationId": "list_voices_api_v1_voices_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/VoiceOption"
         },
         "type": "array",
         "title": "Response List Voices Api V1 Voices Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/agents/default": {
   "get": {
    "tags": [
     "voice"
    ],
    "summary": "The tenant's default AI agent",
    "operationId": "get_default_agent_api_v1_agents_default_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AgentResponse"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   },
   "put": {
    "tags": [
     "voice"
    ],
    "summary": "Configure the default AI agent (owner/admin)",
    "operationId": "update_default_agent_api_v1_agents_default_put",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AgentUpdate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AgentResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/voice/sessions": {
   "post": {
    "tags": [
     "voice"
    ],
    "summary": "Start a browser voice call with the default agent",
    "operationId": "create_voice_session_api_v1_voice_sessions_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "anyOf": [
         {
          "$ref": "#/components/schemas/VoiceSessionRequest"
         },
         {
          "type": "null"
         }
        ],
        "title": "Body"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/VoiceSessionResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/voice/calls/{call_id}/transcript": {
   "get": {
    "tags": [
     "voice"
    ],
    "summary": "What was said (until it expires)",
    "operationId": "get_transcript_api_v1_voice_calls__call_id__transcript_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "call_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Call Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TranscriptResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/voice/calls/{call_id}": {
   "get": {
    "tags": [
     "voice"
    ],
    "summary": "Call status and usage",
    "operationId": "get_call_api_v1_voice_calls__call_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "call_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Call Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CallResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/voice/latency": {
   "get": {
    "tags": [
     "voice"
    ],
    "summary": "Turn latency across the tenant's recent calls",
    "operationId": "get_latency_api_v1_voice_latency_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "days",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 90,
       "minimum": 1,
       "default": 7,
       "title": "Days"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/LatencyResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "AgentResponse": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "company_name": {
      "type": "string",
      "title": "Company Name"
     },
     "company_description": {
      "type": "string",
      "title": "Company Description"
     },
     "objective": {
      "type": "string",
      "title": "Objective"
     },
     "greeting": {
      "type": "string",
      "title": "Greeting"
     },
     "language": {
      "type": "string",
      "title": "Language"
     },
     "voice": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Voice"
     },
     "personality": {
      "type": "string",
      "title": "Personality",
      "default": ""
     },
     "qualification_questions": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Qualification Questions",
      "default": []
     },
     "objection_guidance": {
      "type": "string",
      "title": "Objection Guidance",
      "default": ""
     },
     "escalation_guidance": {
      "type": "string",
      "title": "Escalation Guidance",
      "default": ""
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "company_name",
     "company_description",
     "objective",
     "greeting",
     "language",
     "voice"
    ],
    "title": "AgentResponse"
   },
   "AgentUpdate": {
    "properties": {
     "name": {
      "type": "string",
      "maxLength": 80,
      "minLength": 1,
      "title": "Name"
     },
     "company_name": {
      "type": "string",
      "maxLength": 200,
      "minLength": 1,
      "title": "Company Name"
     },
     "company_description": {
      "type": "string",
      "maxLength": 4000,
      "minLength": 1,
      "title": "Company Description"
     },
     "objective": {
      "type": "string",
      "maxLength": 1000,
      "minLength": 1,
      "title": "Objective"
     },
     "greeting": {
      "type": "string",
      "maxLength": 500,
      "minLength": 1,
      "title": "Greeting"
     },
     "language": {
      "type": "string",
      "pattern": "^[a-z]{2}(-[A-Z]{2})?$",
      "title": "Language",
      "default": "en-IN"
     },
     "voice": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 100,
        "pattern": "^[A-Za-z0-9_.:-]+$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Voice"
     },
     "personality": {
      "type": "string",
      "maxLength": 300,
      "title": "Personality",
      "default": ""
     },
     "qualification_questions": {
      "items": {
       "type": "string",
       "maxLength": 200,
       "minLength": 3
      },
      "type": "array",
      "maxItems": 10,
      "title": "Qualification Questions"
     },
     "objection_guidance": {
      "type": "string",
      "maxLength": 2000,
      "title": "Objection Guidance",
      "default": ""
     },
     "escalation_guidance": {
      "type": "string",
      "maxLength": 1000,
      "title": "Escalation Guidance",
      "default": ""
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "name",
     "company_name",
     "company_description",
     "objective",
     "greeting"
    ],
    "title": "AgentUpdate",
    "description": "Tenant-supplied agent configuration. Bounded so it cannot swamp the prompt."
   },
...
   "CallResponse": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "channel": {
      "type": "string",
      "title": "Channel"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "sales_state": {
      "type": "string",
      "title": "Sales State"
     },
     "turn_count": {
      "type": "integer",
      "title": "Turn Count"
     },
     "started_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Started At"
     },
     "ended_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ended At"
     },
     "end_reason": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "End Reason"
     },
     "usage": {
      "$ref": "#/components/schemas/CallUsage"
     },
     "latency": {
      "$ref": "#/components/schemas/LatencyResponse"
     }
    },
    "type": "object",
    "required": [
     "id",
     "channel",
     "status",
     "sales_state",
     "turn_count",
     "started_at",
     "ended_at",
     "end_reason",
     "usage",
     "latency"
    ],
    "title": "CallResponse"
   },
...
   "CallUsage": {
    "properties": {
     "llm_input_tokens": {
      "type": "integer",
      "title": "Llm Input Tokens"
     },
     "llm_output_tokens": {
      "type": "integer",
      "title": "Llm Output Tokens"
     },
     "llm_turns": {
      "type": "integer",
      "title": "Llm Turns"
     },
     "llm_interrupted_turns": {
      "type": "integer",
      "title": "Llm Interrupted Turns"
     },
     "stt_seconds": {
      "type": "number",
      "title": "Stt Seconds"
     },
     "tts_characters": {
      "type": "integer",
      "title": "Tts Characters"
     }
    },
    "type": "object",
    "required": [
     "llm_input_tokens",
     "llm_output_tokens",
     "llm_turns",
     "llm_interrupted_turns",
     "stt_seconds",
     "tts_characters"
    ],
    "title": "CallUsage"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "LatencyResponse": {
    "properties": {
     "agent_turns": {
      "type": "integer",
      "title": "Agent Turns"
     },
     "measured_turns": {
      "type": "integer",
      "title": "Measured Turns"
     },
     "interrupted_turns": {
      "type": "integer",
      "title": "Interrupted Turns"
     },
     "e2e_p50_ms": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "E2E P50 Ms"
     },
     "e2e_p95_ms": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "E2E P95 Ms"
     },
     "llm_ttft_p50_ms": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Llm Ttft P50 Ms"
     },
     "tts_ttfb_p50_ms": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Tts Ttfb P50 Ms"
     },
     "end_of_turn_delay_p50_ms": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "End Of Turn Delay P50 Ms"
     },
     "target_p50_ms": {
      "type": "integer",
      "title": "Target P50 Ms"
     },
     "target_p95_ms": {
      "type": "integer",
      "title": "Target P95 Ms"
     },
     "meets_target": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Meets Target"
     }
    },
    "type": "object",
    "required": [
     "agent_turns",
     "measured_turns",
     "interrupted_turns",
     "e2e_p50_ms",
     "e2e_p95_ms",
     "llm_ttft_p50_ms",
     "tts_ttfb_p50_ms",
     "end_of_turn_delay_p50_ms",
     "target_p50_ms",
     "target_p95_ms",
     "meets_target"
    ],
    "title": "LatencyResponse",
    "description": "Prospect stops speaking -> agent's first audio, against the Phase 3 targets."
   },
...
   "TranscriptLine": {
    "properties": {
     "seq": {
      "type": "integer",
      "maximum": 100000.0,
      "minimum": 0.0,
      "title": "Seq"
     },
     "speaker": {
      "type": "string",
      "enum": [
       "prospect",
       "agent"
      ],
      "title": "Speaker"
     },
     "text": {
      "type": "string",
      "maxLength": 4000,
      "minLength": 1,
      "title": "Text"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "seq",
     "speaker",
     "text"
    ],
    "title": "TranscriptLine"
   },
   "TranscriptResponse": {
    "properties": {
     "call_id": {
      "type": "string",
      "format": "uuid",
      "title": "Call Id"
     },
     "lines": {
      "items": {
       "$ref": "#/components/schemas/TranscriptLine"
      },
      "type": "array",
      "title": "Lines"
     }
    },
    "type": "object",
    "required": [
     "call_id",
     "lines"
    ],
    "title": "TranscriptResponse"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
   "VoiceOption": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "key": {
      "type": "string",
      "title": "Key"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "gender": {
      "type": "string",
      "enum": [
       "female",
       "male"
      ],
      "title": "Gender"
     },
     "default": {
      "type": "boolean",
      "title": "Default"
     }
    },
    "type": "object",
    "required": [
     "id",
     "key",
     "name",
     "gender",
     "default"
    ],
    "title": "VoiceOption"
   },
   "VoiceSessionRequest": {
    "properties": {
     "lead_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Lead Id"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "title": "VoiceSessionRequest"
   },
   "VoiceSessionResponse": {
    "properties": {
     "call_id": {
      "type": "string",
      "format": "uuid",
      "title": "Call Id"
     },
     "room": {
      "type": "string",
      "title": "Room"
     },
     "livekit_url": {
      "type": "string",
      "title": "Livekit Url"
     },
     "token": {
      "type": "string",
      "title": "Token"
     }
    },
    "type": "object",
    "required": [
     "call_id",
     "room",
     "livekit_url",
     "token"
    ],
    "title": "VoiceSessionResponse"
   }
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 }
}